[core] Extract and de-templatize several Program static methods
